+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - onAuthCodeReceived, developer should call com.ca.mas.foundation.MASUser.login(com.ca.mas.foundation.MASAuthCredentials, com.ca.mas.foundation.MASCallback<com.ca.mas.foundation.MASUser>) with com.ca.mas.foundation.MASAuthCredentialsAuthorizationCode
     - Do not execute pending request when authentication failed. The pending queue will only be executed after authentication success.
 - Refactor MAS interface to resolve Xamarin Binding [US477776]
+- MASUser.getAuthCredentialsType has been removed [DE354252]
 
 ### Deprecated Classes
 - MASOtpAuthFragment.java is removed, please use MASOtpActivity/MASOtpDialogFragment.